+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 - The `num_channels` and `num_frames` arguments have been removed from `Model::process_interleaved`'s function signature. These arguments are now inferred from the buffer size and the `num_channels` value passed to `Model::initialize`.
 - `ModelType::QuailSTT` was renamed to `ModelType::QuailSttL16`.
+- `ModelType::QuailXS` was renamed to `ModelType::QuailXs`.
+- `ModelType::QuailXXS` was renamed to `ModelType::QuailXxs`.
 - `Model::create_vad` signature changed: the model reference now has to be mutable (`&mut self` instead of `&self`).
 
 ### Fixes